feat(product): added update and delete product

+++ internal/product/entity/product.go
@@ -5,7 +5,7 @@ import "gorm.io/gorm"
 type Product struct {
 	gorm.Model
 	Name        string  `gorm:"type:varchar(255);not null"`
-	Price       float64 `gorm:"type:integer;not null"`
+	Price       int `gorm:"type:integer;not null"`
 }
 
 func (Product) TableName() string {
